Update es-419
This fixes the intonation in Latin American Spanish questions, removes the phoneme replacement (T to s) because it is already included in phonemes es_la. It also eliminates the line languaje es 6, because when the voice is in the dialect of Spain and it automatically changes to the dialect of Latin America, it does not return to the dialect of Spain.