Android: Cache successfully found voices in case of partial SSML from the system
Sometimes, `TtsService.findVoice` is called with a language, but without a country and variant. This causes unexpected voice changes (for instance, the user choosing the UK variant and having either the US or RP voice selected seemingly arbitrarily). I'm not sure exactly why this happens, but suspect it's due to incomplete SSML being provided by clients and/or the system.

This commit adds a new "variant cache": whenever a voice with a particular language and country and/or variant is requested, the found voice is stored and, if a subsequent call for the same language doesn't specify a country or variant, the last-known voice is looked up from the cache.

I've tested that changing languages and variants works as expected, and can no longer reproduce the erratic voice changes!

Closes #970.