Android: Set a default English voice
The language code "en" doesn't map onto an eSpeak voice, but Android clients/the system often use "eng". This causes erratic behaviour: in some cases, the RP voice is used. In others, the US voice is used. Since the standard UK English voice is eSpeak's original English voice, maps onto "en" when using the command line implementation, and is (in my opinion) eSpeak's highest quality voice by far, this commit explicitly makes GBR the default country for English when none is specified.

This doesn't solve the problem in any other languages where it might exist. I tried a more generalized approach in #1885, but there are problems when reading web content with inline variant changes within a language. A more general approach should be considered, but in the meantime, this should be merged soon as it provides significant quality of life improvements in eSpeak's primary language.

Closes #970.