Add a list of names with the correct pronunciations in Portuguese
The names in the list are not pronounced correctly by espeak-ng when
they are without accents/diacritics. For example, Mario is pronounced
[maɾˈiʊ] instead of [mˈaɾjʊ].

As we know, many systems store people's names without accents due to
technical reasons. Additionally, many people use their names without an
accent online. The update aims to improve espeak-ng's support for these
technical and cultural facts.

The main reason for the addition is that speech synthesis is also used
to call people waiting for healthcare by name.