Fix the replacement character in the input string
The correction uses substitute character (0x001A) as a workaround. This
control character is the "older brother" of the replacement character
OxFFFD.

The workaround is necessary because the replacement character (0xFFFD)
is used to flag to indicate an error or EOF. The substitute character
(0x001A) was included as an alternative flag to indicate that the input
character itself is a replacement character (0xFFFD).