Fix the "I umlaut a half" bug
It fixes the bug that can be listened as "I umlaut a half", when there's
a replacement character (0xFFFD) in the input text.

The function that decodes UTF-8 characters was changed to return a
substitute character (0x001A) when it detects a replacement character
(0xFFFD) in the input text.

This workaround is necessary because the replacement character (0xFFFD)
is used internally to indicate an invalid character, an internal error
or an end of file.