[hebrew] Treat final apostrophe as part of the word
Set LOPT_APOSTROPHE bit 1 for Hebrew to ensure final apostrophes are treated as part of the word. This is relevant for words like בָּגָאז׳ or חָגָ׳ג׳, where the apostrophe represents a phonemic element and should not be ignored during tokenization or phoneme conversion.

Replacement rules will change  ׳ (Hebrew GERESH = U+05F3) to ' (standard apostrophe)