Mark Truncating Casts as Intentional
We see false positives in run-time checks of truncating casts (`/RTCc` on MSVC). The loss of bits is intentional here and this commit marks it so.

The additional bitmasks are optimized away by every optimizing compiler, but they remove extra bits before the cast to `char` takes place so no implicit data loss occurs.