pt: change unstressed /i/ phonemes
Updated phonemes:

*  [y]: a wildcard for unstressed /i/ in Brazil and unstressed /i/ in
Portugal.

*  [I]: is now an unstressed /i/ for Brazilian dialect. Previously it
was the epenthetic /i/.

*  [I-]: a very short /i/ to be inserted in between consontants to
break uncommon consonant clusters in Brazilian dialect, also known as
epenthetic /i/.

The epenthetic vowels in `pt_rule` were also updated to match the new
[I-] phoneme.

It only affects the voices that use the `?3` conditional, which are
these mbrola voices: mb-br1, mb-br2, mb-br3 and mb-br4. The default
`pt-br` (robotic) voice is not affected.