pt: remove english linking /r/ from pt phoneme table
The "en/r-" serves as a linking /r/, which occurs in English expressions
like "law and order": law**r** and order.

Also, when "en/r-" is at the end of the word, it is redirected to
"base1/r".

In the "pt_list" file, it was used at the end of words. Therefore, it is
more appropriate to use "base1/r" directly.

Linking R: https://en.wikipedia.org/wiki/Linking_and_intrusive_R