pt: Adjust alveolar tap in complex onset
Now the short schwa will be transparently inserted into the complex
onset formed by a consoant and an alveolar tap.

NOTE:
In Brazilian Portuguese, the "supporting vowel" in a complex onset is
not a schwa. It's actually a very short vowel, similar to the vowel
following the alveolar tap. This is an improvement we'll consider for
the future.