pt: update the spacial phoneme for /s/
Update the old spacial phoneme [s#], which is used for /s/ in coda.

Also update the new special phonemes [r#] and [R#], weak /r/ and strong
/R/, respectively.

Now the special phonemes are described as `virtual` and they are
redirected by the instruction ChangePhoneme() to the target phoneme.

The `ChangePhoneme()` instruction also changes the characteres shown
with the `-x` option, so that [s] is printed  in the phonetic
transcription instead of [s#].