pt: make use of the special phonemes for strong /R/ and weak /r/
Update `pt_rules` file to use the special phonemes for strong /R/ and
weak /r/. They allow the inclusion of more dialects in the future.

The old phonemes [*] and [x] will be preserved for backward
compatibility.

NOTE:
There are exceptions in `pt_rules` and `pt_list` files that still use
the [*] phoneme. They should be updated when another dialect is included
in espeak-ng.