pt: fix mbrola pronunciation of the open-mid back rounded vowel
This commit removes the line that maps Kirshenbaum [O] to [oo].

There's no [oo] phoneme in the `pt1` mbrola voice. When we try to
translate a word with this vowel, a warning is shown and the diphones
are replaced with a silence.

	espeak-ng -v mb-pt1 "porta" -x
	p'O*@-t&
	mbrola: Warning: p-oo unknown, replaced with _-_
	mbrola: Warning: oo-r unknown, replaced with _-_

Probably the `pt1` file was based on the `ptbr` file.