pt: fix mbrola pronunciation of the diphthong "éu"
This commit changes the line that maps Kirshenbaum [EU] to [eew].

There's no [ee] phoneme in the `pt1` mbrola voice. When we try to
translate a word with this vowel, a warning is shown and the diphones
are replaced with a silence.

	$ espeak-ng -v mb-pt1 "céus" -x
	s'EUS
	mbrola: Warning: s-ee unknown, replaced with _-_
	mbrola: Warning: ee-w unknown, replaced with _-_

Probably the `pt1` file was based on the `ptbr` file.