pt: add phoneme description for the [oU] diphthong
The old diphthong formed by [o] and [w] was very long and often
negatively affected the prosody of the sentence, making it sound
strange.

Also updated the dictionary files and mbrola files.

NOTE: to avoid hiatus, a [w] is appended when the next sound is a vowel.
The same thing is done in some other diphthongs to avoid hiatus.