pt: remove intervocalic alternative phonemes for /b, d, g/
Unfortunately, there is no easy way to check whether a vowel is nasal or
nasalized.

The condiction `prevPhW(isNasal)` doesn't work because the `nas` (nasal)
flag indicates the phoneme is a consonant, not a vowel.

Also, the `nzd` (nasalized) flag is not supported by espeak-ng, although
it's in the documentation.