pt: update tests for Portuguese
Some phonemes have changed.

The sound of the phoneme [l^] has been unified in both Brazilian and
European Portuguese.

The phoneme [R] has been fixed to sound like a liquid trill.
Previously, it sounded like an English connecting R in European
Portuguese.

The phoneme [*] now is helped by a very short schwa [@-] when in complex
onset or internal coda.

The phoneme [s#] is now a virtual phoneme that switches between [s] or
[z] and [S] or [Z] depending on dialect and context.

Also updates a test for Xextan because this constructed language it is
based on Brazilian Portuguese.