pt: change the epenthetic /i/
The main reason for this change is to use the slash in the phoneme name,
aka variant indicator, instead of the dash, which has a specific meaning
in the Kirshenbaum system. Another reason is to keep the phoneme
mnemonics in the standard output backward compatible.

The epenthetic /i/ is a very short vowel to be inserted in between
consontants to break uncommon consonant clusters in Brazilian dialect.

Currently, it is used in Brazilian mbrola voices as they have an
incomplete set of diphones.