Fix Sindarin medial preconsonantal F
According to Appendix E of The Lord of the Rings, ⟨f⟩ represents /f/
except at the end of words, and that’s what I previously implemented
here. However, according to the later essay “The Rivers and Beacon-hills
of Gondor”, published in The Nature of Middle-earth pp. 376–396
(relevant passage on p. 386; the essay was also published in Vinyar
Tengwar issue 42, which I do not have access to), voiceless /f/ did not
occur medially before consonants in Sindarin, and voiced /v/ may
therefore be spelled by ⟨f⟩ in this position (⟨v⟩ before consonants
being avoided in English). Implement this with two new tests: Lefnui,
the river name that prompted Tolkien to note this spelling peculiarity
(he spelled it Levnui in the essay, but noted that it was spelled Lefnui
on the Map and that this spelling was preferable), and Fladrif, an Ent
name that demonstrates that initial ⟨f⟩ before a consonant is still
pronounced /f/.